fix: translate in MainPopUp, moved ReviewPopUp
