+++ client/src/components/form-components/Input/Input.jsx
@@ -1,9 +1,10 @@
+import Heading from '@components/UI/Texts/Heading/Heading';
 import cl from './index.module.scss';
 
 export function Input({ labelText, type = 'text', name, variant, register, errors, ...rest }) {
   return (
     <label htmlFor={name}>
-      {labelText}
+      <Heading type="h4">{labelText}</Heading>
 
       <input id={name} className={`${cl.input} ${cl[variant]}`} {...register(name)} type={type} {...rest} />
