+++ client/src/components/form-components/Textarea/Textarea.jsx
@@ -1,9 +1,12 @@
+import Heading from '@components/UI/Texts/Heading/Heading';
 import cl from './index.module.scss';
 
 export function Textarea({ labelText, errors, name, variant, register, ...rest }) {
   return (
     <>
-      <label htmlFor={name}>{labelText}</label>
+      <label htmlFor={name}>
+        <Heading type="h4">{labelText}</Heading>
+      </label>
       <textarea className={`${cl.textarea} ${cl[variant]}`} {...register(name)} {...rest} id={name} />
       {errors && errors[name] && <p style={{ color: 'red' }}>{errors[name].message}</p>}
     </>
